bulk-email-sender/backend/requirements.txt
# Core email functionality
python-dotenv==1.0.0
aiosmtplib==3.0.1
email-validator==2.1.0

# Web framework
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Task queue
celery==5.3.4
redis==5.0.1

# Email templates
jinja2==3.1.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring and tracking
prometheus-client==0.19.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0

# Email authentication
dkimpy==1.1.5
dnspython==2.4.2

# Rate limiting
slowapi==0.1.9

# CSV processing
pandas==2.1.4
openpyxl==3.1.2